Handle documentation for different targets
When cargo is instructed to build code for a different target, it will store the
output under `target/<host-triple>`. This means that some docs will be under
`target/<triple>/doc` and some under `target/doc`.

This change takes into the account the existence of `CARGO_BUILD_TARGET` and
uses that to discover newly created docs when copying/processing.